update popular item and add map

+++ lib/feature/popular_items/models/popular_item.dart
@@ -1,8 +1,7 @@
 class PopularItem {
   final String text;
   final String image;
-  final int price;
-
+  final double price;
   PopularItem({
     required this.text,
     required this.image,
